+++ tests/unit/tile/test_tile_component_manager.py
@@ -2432,6 +2432,75 @@ class TestStaticSimulator:  # pylint: disable=too-many-public-methods
         )
         tile_component_manager.configure_test_generator(**mocked_input_params)
 
+    def test_configure_pattern_generator(
+        self: TestStaticSimulator,
+        tile_component_manager: TileComponentManager,
+        tile_simulator: TileSimulator,
+    ) -> None:
+        """
+        Unit test for the configure_pattern_generator function.
+
+        :param tile_component_manager: The TileComponentManager instance.
+        :param tile_simulator: The tile simulator instance.
+        """
+        tile_simulator.connect()
+        mocked_input_params: dict[str, Any] = {
+            "stage": "jesd",
+            "pattern": list(range(1024)),
+            "adders": list(range(32)),
+            "start": True,
+            "shift": 0,
+            "zero": 0,
+        }
+
+        tile_simulator.set_pattern = unittest.mock.Mock()  # type: ignore[assignment]
+
+        tile_component_manager.configure_pattern_generator(**mocked_input_params)
+        tile_simulator.set_pattern.assert_called_with(
+            mocked_input_params["stage"],
+            mocked_input_params["pattern"],
+            mocked_input_params["adders"],
+            mocked_input_params["start"],
+            mocked_input_params["shift"],
+            mocked_input_params["zero"],
+        )
+
+    def test_start_pattern_generator(
+        self: TestStaticSimulator,
+        tile_component_manager: TileComponentManager,
+        tile_simulator: TileSimulator,
+    ) -> None:
+        """
+        Unit test for the start_pattern_generator function.
+
+        :param tile_component_manager: The TileComponentManager instance.
+        :param tile_simulator: The tile simulator instance.
+        """
+        tile_simulator.connect()
+
+        mocked_stage = "jesd"
+        tile_simulator.start_pattern = unittest.mock.Mock()  # type: ignore[assignment]
+        tile_component_manager.start_pattern_generator(mocked_stage)
+        tile_simulator.start_pattern.assert_called_with(mocked_stage)
+
+    def test_stop_pattern_generator(
+        self: TestStaticSimulator,
+        tile_component_manager: TileComponentManager,
+        tile_simulator: TileSimulator,
+    ) -> None:
+        """
+        Unit test for the stop_pattern_generator function.
+
+        :param tile_component_manager: The TileComponentManager instance.
+        :param tile_simulator: The tile simulator instance.
+        """
+        tile_simulator.connect()
+
+        mocked_stage = "jesd"
+        tile_simulator.stop_pattern = unittest.mock.Mock()  # type: ignore[assignment]
+        tile_component_manager.stop_pattern_generator(mocked_stage)
+        tile_simulator.stop_pattern.assert_called_with(mocked_stage)
+
     def test_test_generator_input_select(
         self: TestStaticSimulator,
         tile_component_manager: TileComponentManager,