+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version History
 
+* [MCCS-2300] Add support for starting/stopping/configuring the pattern generator in MccsTile.
 * [MCCS-2278] Change health and state rollup/aggregation for SpsStation to reduce frequency of it entering `DevState.UNKNOWN`.
 
 ## 0.17.7